% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFAtree_functions.R
\name{EFAtree}
\alias{EFAtree}
\title{Function to grow an EFA tree}
\usage{
EFAtree(data, covariates, model, alpha = 0.05, maxdepth = 3, minsize = 100)
}
\arguments{
\item{data}{Data frame containing only columns of the observed variables (items) in EFA model.}

\item{covariates}{Data frame containing only columns of the covariates to be tested by the tree.}

\item{model}{EFA model in lavaan syntax.}

\item{alpha}{Level of significance to be used in tree. Defaults to 0.05.}

\item{maxdepth}{Maximum depth of tree, including the parent node. Defaults to 3.}

\item{minsize}{Minimum number of observations in each node. Defaults to 100.}
}
\description{
Function to grow an EFA tree. Simultaneously tests parameter stability in an EFA model across multiple (categorical or continuous) covariates
(i.e., tests of measurement invariance). For more details, see Sterner and Goretzko (2023) or Zeileis et al. (2008).
}
