% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFAtree_helpers.R
\name{EFAtree_plot}
\alias{EFAtree_plot}
\title{Functions to plot an EFA tree and inspect its notes}
\usage{
EFAtree_plot(treeobj)
}
\arguments{
\item{treeobj}{Fitted EFA tree model.}

\item{node.id}{ID to identify a node in the tree.}
}
\value{
.
}
\description{
Functions to plot an EFA tree and inspect its notes
}
\references{
Sterner, P., & Goretzko, D. (2023). Exploratory factor analysis trees: Evaluating measurement invariance between multiple covariates. Structural Equation Modeling: A Multidisciplinary Journal, 30, 871–886. https://doi.org/10.1080/10705511.2023.2188573

Zeileis, A., Hothorn, T., & Hornik, K. (2008). Model-based recursive partitioning. Journal of Computational and Graphical Statistics, 17, 492–514. https://doi.org/10.1198/106186008X319331
}
