% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGAtree_functions.R
\name{EGAtree}
\alias{EGAtree}
\title{Function to grow an EGA tree}
\usage{
EGAtree(data, covariates, ...)
}
\arguments{
\item{data}{Data frame containing only columns of the observed variables (items).}

\item{covariates}{Data frame containing only columns of the covariates to be tested by the tree.}

\item{...}{additional arguments passed to \link{mob_control} (mob) or \link{ctree_control} (ctree).}
}
\value{
A list containing: an object of class \code{mob} and a list as returned by \code{\link[EGAnet:EGA]{EGAnet::EGA()}}.
}
\description{
Function to grow an EFA tree. Simultaneously tests parameter stability in an EFA model across multiple (categorical or continuous) covariates
(i.e., tests of measurement invariance).
}
\references{
Goretzko, D., & Sterner, P. (2024). Exploratory Graph Analysis Trees - A Network-based Approach to Investigate Measurement Invariance with Numerous Covariates. https://doi.org/10.31234/osf.io/9cx8z

Jones, P.J., Mair, P., Simon, T., & Zeileis, A. (2020) Network Trees: A Method for Recursively Partitioning Covariance Structures. Psychometrika, 85, 926–945. https://doi.org/10.1007/s11336-020-09731-4

Golino, H., & Christensen, A. P. (2024). EGAnet: Exploratory Graph Analysis – A framework for estimating the number of dimensions in multivariate data using network psychometrics. https://doi.org/10.32614/CRAN.package.EGAnet
}
