% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGAtree_helpers.R
\name{EGAtree_teststats}
\alias{EGAtree_teststats}
\title{Functions to plot an EGA tree and inspect its notes}
\usage{
EGAtree_teststats(treeobj, node.id = NULL)
}
\arguments{
\item{treeobj}{Fitted EFA tree model.}

\item{node.id}{ID to identify a node in the tree.}
}
\value{
.
}
\description{
Functions to plot an EGA tree and inspect its notes
}
